backend/requirements.txt
Django~=4.2.4
channels[daphne]~=3.0.4
djangorestframework~=3.14.0
coreapi~=2.3.3